temp sensor now reads from analog pin. Test display changes as well as test monitor core changes

+++ src/display/TestMonitorDisplay.h
@@ -17,7 +17,7 @@ class TestMonitorDisplay : public MonitorDisplay {
         void operator=(TestMonitorDisplay const&) = delete;
 
         void print(int x, int y, const char *text, Color color);
-        void printSensorScreen(int tempSensorValue, int oilSensorValue);
+        void printSensorScreen(float tempSensorValue, float oilSensorValue);
 
     private:
         TestMonitorDisplay();
